refactor: 拆分nowplaying通知订阅事件
